Linux: fix various compilation issues under Linux.

+++ src/Main/TextUserInterface.cpp
@@ -640,7 +640,7 @@ namespace VeraCrypt
 				}
 				else if (sizeStr.find (L"T") != string::npos)
 				{
-					multiplier = 1024 * 1024 * 1024 * 1024;
+					multiplier = (uint64) 1024 * 1024 * 1024 * 1024;
 					sizeStr.resize (sizeStr.size() - 1);
 				}
